+++ src/layouts/ArticleHead.vue
@@ -45,6 +45,9 @@
             }}<span v-if="index !== frontmatter.author.length - 1">, </span>
           </template>
         </span>
+        <span>
+            &middot; {{ readingTime }} minutes to read
+        </span>
         <span class="tags">
           &middot;
           <span v-for="(tag, index) in frontmatter.tags" :key="index">{{
@@ -63,8 +66,25 @@
 
 <script setup lang="ts">
 import { useData } from "vitepress";
+import { computed, onMounted, ref } from "vue";
+
 const { frontmatter, theme } = useData();
+
+const contentRef = ref<HTMLElement | null>(null);
+const readingTime = ref<string>('');
+
+onMounted(() => {
+  contentRef.value = document.getElementById('content');
+  if (contentRef.value) {
+    const words = contentRef.value.innerText.trim().split(/\s+/).length;
+    const wordsPerMinute = 200; // Average reading speed
+    const time = Math.ceil(words / wordsPerMinute);
+    readingTime.value = time.toString();
+  }
+});
+
 </script>
+
 
 <style lang="scss">
 #article-head {